V0.2.0

+++ Backend/src/main/kotlin/org/pin/backend/dto/Data/BorradorGastoDTO.kt
@@ -1,7 +1,9 @@
 package org.pin.backend.dto.Data
 
+import java.math.BigDecimal
+
 data class BorradorGastoDTO(
     val concepto: String,
-    val total: Double,
+    val total: BigDecimal,
     val urlTicket: String,
 )
