+++ Backend/src/main/kotlin/org/pin/backend/dto/Request/GastoRequestDTO.kt
@@ -1,9 +1,11 @@
 package org.pin.backend.dto.Request
 
+import java.math.BigDecimal
+
 data class GastoRequestDTO(
     val nombre: String,
     val descripcion: String?,
-    val importe: Double,
+    val importe: BigDecimal,
     val categoria: String,
     val pagadoPorId: Long,
     val beneficiarios: List<String>? = emptyList(),
