+++ Backend/src/main/kotlin/org/pin/backend/dto/Response/GastoResponseDTO.kt
@@ -1,10 +1,12 @@
 package org.pin.backend.dto.Response
 
+import java.math.BigDecimal
+
 data class GastoResponseDTO(
     val id: Long,
     val nombre: String,
     val descripcion: String?,
-    val importe: Double,
+    val importe: BigDecimal,
     val fecha: String,
     val categoria: String,
     val pagadoPorNombre: String,
